Better naming for routes and files
This will allow for future expansion if needed.
